Fix multi-version latest tagging (#555)
* Fix old-version github query+improve --dry logging

1. github query incorrectly used the *name*, which has spaces for old
versions like expo/v26. Now it uses subdirectory name.
2. For publish-packages, all dry-run ifs are now if/elses that log even
in the dry-run case so you can see what is skipped.

* Add debugging file for VS Code

* Thread correct logger through

instead of using's node superior, but not customisable, `log`.

* Add debug launch for calculate-versions

* Calculate current version first

This makes us publish the current version first as well. This is
important because when we publish older versions, npm moves the "latest"
tag to the older version. So we have to reset it to the current version.
But if a PR updates both current and older versions, we need to make
sure to publish the current version first so that older versions use the
correct "latest" target.

Before:

Publish expo v27 (24.0.14) and current (31.0.9):
1. Publish expo v27 -- 24.0.14
2. Move "latest" tag back to expo current (31.0.9).
ERROR: npm only has 31.0.8, the previous version of current.

After

Publish expo (31.0.9) and expo v27 (24.0.14):
1. Publish expo -- 31.0.9
2. Publish expo v27 -- 24.0.14
3. Move "latest" tag back to expo current (31.0.9).