Check deprecated package dependents (#583)
* Still printf debugging

* Oh, it's the dependents of the package

* Working version

I am considering how to write tests, but not sure yet if it's going to
be worthwhile.

This version is extremely messy and needs lots of cleanup.

* Some cleanup

* More cleanup

* More cleanup

* Some more cleanup

* Finally done with cleanup

I want to come up with tests, but it may be too onerous.

* Disable e2e test and refactor tests

* Add getAffectedPackages test

Doesn't pass, and has a surprising side-effect on the file system.

* Add delete affected package test and refactor